move user reports logic to utils
